Refactored components and screens folders
