+++ training-react/src/app/components/App/components/Field/index.js
