+++ training-react/src/app/components/App/components/Loading/layout.js
