+++ training-react/src/app/components/App/components/Loading/styles.scss
