+++ training-react/src/app/components/App/components/Topbar/styles.scss
