+++ training-react/src/app/components/App/screens/App/components/Game/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import Board from '../Board';
+import Board from '../../screens/Board';
 
 import calculateWinner from './utils';
 import styles from './styles.scss';
