+++ training-react/src/app/components/App/screens/App/components/Game/styles.scss
