+++ training-react/src/app/components/App/screens/App/components/User/layout.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-import withLoadingScreen from '../LoadingScreen';
-import UserInfo from '../UserInfo';
-import UserEdit from '../UserEdit';
+import withLoadingScreen from '../../../../components/Loading';
+import UserInfo from '../../screens/UserInfo';
+import UserEdit from '../../screens/UserEdit';
 
 import styles from './styles.scss';
