+++ training-react/src/app/components/App/screens/App/components/User/styles.scss
