+++ training-react/src/app/components/App/screens/App/screens/Board/index.js
