+++ training-react/src/app/components/App/screens/App/screens/Board/styles.scss
