+++ training-react/src/app/components/App/screens/App/screens/Square/index.js
