+++ training-react/src/app/components/App/screens/App/screens/UserInfo/styles.scss
